COMP: Fixed warning on type conversion on linux
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=4413
git-svn-id: http://svn.slicer.org/Slicer4/trunk@4413 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/BatchMakeApplications/GaussianBlurBatchMakeModule/GaussianBlurBatchMakeModule.cxx
@@ -168,7 +168,7 @@ int main(int argc, char* argv[])
   bm::ProgressManagerSlicer progressManager;
   
   // Come up with an approximation of the number of Run()
-  unsigned int actions = (sigmaMax-sigmaMin)/sigmaStep;
+  unsigned int actions = (int)((sigmaMax-sigmaMin)/sigmaStep);
   if(actions == 0)
     {
     actions = 1;
